[upd] - reorganize livewire directory folder
