+++ resources/views/dashboard/transactions.blade.php
@@ -2,7 +2,7 @@
     <x-dashboard-layout>
         <main class="min-h-[70vh] px-4 pt-6 max-w-[90rem] mx-auto">
             <h1 class="mb-5 text-3xl font-bold">Transactions</h1>
-            <livewire:transaction-dashboard-table>
+            <livewire:dashboard.transaction-dashboard-table>
         </main>
     </x-dashboard-layout>
 </x-layout>
