+++ resources/views/dashboard/users.blade.php
@@ -2,7 +2,7 @@
     <x-dashboard-layout>
         <main class="min-h-[70vh] px-4 pt-6 max-w-[90rem] mx-auto">
             <h1 class="mb-5 text-3xl font-bold">Users</h1>
-            <livewire:users-dashboard-table>
+            <livewire:dashboard.users-dashboard-table>
         </main>
     </x-dashboard-layout>
 </x-layout>
